Renamed method, getErrorMessage unit tests
