+++ src/firebot/effects/spotifyChangePlaybackStateEffect.ts
@@ -1,6 +1,6 @@
 import { spotify } from "@/main";
 import { Firebot } from "@crowbartools/firebot-custom-scripts-types";
-import { getErrorMessage } from "@utils/errors";
+import { getErrorMessage } from "@utils/strings";
 
 export enum SpotifyPlaybackState {
   Play = "Play",
