--- src/utils/errors.ts
@@ -1,2 +1,0 @@
-export const getErrorMessage = (error: any): string =>
-  error instanceof Error ? error.message : "Unhandled Exception";
