+++ src/utils/strings.ts
@@ -1,4 +1,4 @@
-export function msToFormattedString(ms: number, showHours: boolean): string {
+export function formatMsToTimecode(ms: number, showHours: boolean): string {
   const normalizedSeconds = ~~(ms / 1000);
   const hours = ~~(normalizedSeconds / 3600);
   const minutes = ~~((normalizedSeconds / 60) % 60);
@@ -7,3 +7,6 @@ export function msToFormattedString(ms: number, showHours: boolean): string {
     (hours || showHours) && minutes < 10 ? "0" : ""
   }${minutes}:${seconds < 10 ? "0" : ""}${seconds}`;
 }
+
+export const getErrorMessage = (error: any): string =>
+  error instanceof Error ? error.message : "Unhandled Exception";
